[bar-n1duqk] Добавил раздельные страницы регистрации и авторизации. Протестировал.
