+++ src/container/container.html
@@ -87,8 +87,8 @@
           <!-- Исполнитель. -->
           <div id="executor" class="tabcontent">
             <span class="text-role">
-              Получай задания с гарантией
-              от заказчиков и зарабатывай
+              Получайте задания с гарантией
+              от заказчиков и зарабатывайте
               на постоянном потоке заданий.
             </span>
 
@@ -465,7 +465,7 @@
   </div>
 
     <div class="col-lg-12 d-flex justify-content-center pt-3 pb-3">
-      <button class="btn-create-task form-control-sm btn-create-task btn-all-tasks mb-3">
+      <button class="btn-create-task form-control-sm btn-all-tasks mb-3">
         ВСЕ ЗАДАНИЯ
       </button>
     </div>
@@ -550,7 +550,7 @@
 
       <div class="col-lg-12 text-center mb-5">
         <p class="content-title">
-          Выполняйте задания, повышайте свой рейтинг, нарабатывайте положительные отзывы и зарабатывайте деньги.
+          Выполняйте задания, повышайте свой рейтинг, нарабатывайте положительные отзывы и зарабатывайтете деньги.
         </p>
       </div>
     </div>
