+++ src/container/container.js
@@ -27,8 +27,8 @@ export default {
             aAdvantages: [],
             aProveliges: [],
             sPassword: null,
-            aHope: [],
-            role: null    
+            aHope: []
+            // role: null    
         }
     },
     created() {
@@ -124,53 +124,55 @@ export default {
             }
         },
 
-        onChangeTab(evt, type) {
-            var i, tabcontent, tablinks;
+        // TODO: Перенесено на отдельную страницу.
+        // onChangeTab(evt, type) {
+        //     var i, tabcontent, tablinks;
 
-            if (type == "login") {
-                $(".tab-role").removeClass("role-show");
-                $(".tab-role").addClass("role-hide");
-                $(".register").removeClass("selected-role");
-                $(".tab-role").addClass("not-selected-role");
-            }
+        //     if (type == "login") {
+        //         $(".tab-role").removeClass("role-show");
+        //         $(".tab-role").addClass("role-hide");
+        //         $(".register").removeClass("selected-role");
+        //         $(".tab-role").addClass("not-selected-role");
+        //     }
 
-            if (type == "register") {
-                $(".tab-role").removeClass("role-hide");
-                $(".tab-role").addClass("role-show");
-                $(".register").addClass("selected-role");
-            }
+        //     if (type == "register") {
+        //         $(".tab-role").removeClass("role-hide");
+        //         $(".tab-role").addClass("role-show");
+        //         $(".register").addClass("selected-role");
+        //     }
 
-            if (type == "executor") {
-                this.role = "E";
-            }
+        //     if (type == "executor") {
+        //         this.role = "E";
+        //     }
 
-            if (type == "customer") {
-                this.role = "C";
-            }
+        //     if (type == "customer") {
+        //         this.role = "C";
+        //     }
 
-            // Get all elements with class="tabcontent" and hide them
-            tabcontent = document.getElementsByClassName("tabcontent");
+        //     // Get all elements with class="tabcontent" and hide them
+        //     tabcontent = document.getElementsByClassName("tabcontent");
 
-            for (i = 0; i < tabcontent.length; i++) {
-                tabcontent[i].style.display = "none";
-            }
+        //     for (i = 0; i < tabcontent.length; i++) {
+        //         tabcontent[i].style.display = "none";
+        //     }
 
-            // Get all elements with class="tablinks" and remove the class "active"
-            tablinks = document.getElementsByClassName("tablinks");
+        //     // Get all elements with class="tablinks" and remove the class "active"
+        //     tablinks = document.getElementsByClassName("tablinks");
 
-            for (i = 0; i < tablinks.length; i++) {
-                tablinks[i].className = tablinks[i].className.replace(" active", "");
-            }
+        //     for (i = 0; i < tablinks.length; i++) {
+        //         tablinks[i].className = tablinks[i].className.replace(" active", "");
+        //     }
 
-            // Show the current tab, and add an "active" class to the button that opened the tab
-            document.getElementById(type).style.display = "block";
-            evt.currentTarget.className += " active";
-        },
+        //     // Show the current tab, and add an "active" class to the button that opened the tab
+        //     document.getElementById(type).style.display = "block";
+        //     evt.currentTarget.className += " active";
+        // },
 
-        // Проставит по дефолту выделенный таб.
-        setSelectedRole() {
-            document.getElementById("defaultOpen").click();
-        },
+        // TODO: Перенесено на отдельную страницу.
+        // // Проставит по дефолту выделенный таб.
+        // setSelectedRole() {
+        //     document.getElementById("defaultOpen").click();
+        // },
 
         // Функция выгружает данные для блока "Наши преимущества".
         _loadPriveleges() {
@@ -191,76 +193,78 @@ export default {
             }
         },
 
-        // Функция регистрирует юзера.
-        onCreate() {
-            const sUrl = this.oData.urlApi.concat("/user/create");
-            let oData = {};
+        // TODO: Перенесено на отдельную страницу.
+        // // Функция регистрирует юзера.
+        // onCreate() {
+        //     const sUrl = this.oData.urlApi.concat("/user/create");
+        //     let oData = {};
 
-            if (this.role == "C") {
-                oData = {
-                    UserName: $("#idLog").val(),
-                    UserPassword: $("#idPass1").val(),
-                    Email: $("#idEma1").val(),
-                    UserRole: this.role
-                };
-            }
+        //     if (this.role == "C") {
+        //         oData = {
+        //             UserName: $("#idLog").val(),
+        //             UserPassword: $("#idPass1").val(),
+        //             Email: $("#idEma1").val(),
+        //             UserRole: this.role
+        //         };
+        //     }
 
-            else if (this.role == "E") {
-                oData = {
-                    UserName: $("#idLogin").val(),
-                    UserPassword: $("#idPassword").val(),
-                    Email: $("#idEmail").val(),
-                    UserRole: this.role
-                };
-            }            
+        //     else if (this.role == "E") {
+        //         oData = {
+        //             UserName: $("#idLogin").val(),
+        //             UserPassword: $("#idPassword").val(),
+        //             Email: $("#idEmail").val(),
+        //             UserRole: this.role
+        //         };
+        //     }            
 
-            try {
-                axios.post(sUrl, oData)
-                    .then((response) => {
-                        console.log("Регистрация успешна", response);
-                    })
+        //     try {
+        //         axios.post(sUrl, oData)
+        //             .then((response) => {
+        //                 console.log("Регистрация успешна", response);
+        //             })
 
-                    .catch((XMLHttpRequest) => {
-                        throw new Error('Ошибка регистрации', XMLHttpRequest.response.data);
-                    });
-            } 
+        //             .catch((XMLHttpRequest) => {
+        //                 throw new Error('Ошибка регистрации', XMLHttpRequest.response.data);
+        //             });
+        //     } 
             
-            catch (ex) {
-                throw new Error("Неожиданная ошибка", ex);
-            }
-        },
+        //     catch (ex) {
+        //         throw new Error("Неожиданная ошибка", ex);
+        //     }
+        // },
 
-        // Функция авторизует юзера.
-        onLogin() {
-            const sUrl = this.oData.urlApi.concat("/user/login");
-            let oData = {
-                UserName: $("#idEma").val(),
-                UserPassword: this.sPassword
-            };
+        // TODO: Перенесено на отдельную страницу.
+        // // Функция авторизует юзера.
+        // onLogin() {
+        //     const sUrl = this.oData.urlApi.concat("/user/login");
+        //     let oData = {
+        //         UserName: $("#idEma").val(),
+        //         UserPassword: this.sPassword
+        //     };
 
-            try {
-                axios.post(sUrl, oData)
-                    .then((response) => {
-                        // Если токен есть, то в зависимости от роли распределяет по интерфейсам.
-                        if (response.data.userToken) {
-                            sessionStorage["userToken"] = response.data.userToken;
-                            sessionStorage["role"] = response.data.role;
-                            sessionStorage["user"] = response.data.user;
-                            $(".right-panel").hide();
-                            // this.$router.push("/home");
-                            window.location.href = window.location.href.concat("home");
-                        }                        
-                    })
+        //     try {
+        //         axios.post(sUrl, oData)
+        //             .then((response) => {
+        //                 // Если токен есть, то в зависимости от роли распределяет по интерфейсам.
+        //                 if (response.data.userToken) {
+        //                     sessionStorage["userToken"] = response.data.userToken;
+        //                     sessionStorage["role"] = response.data.role;
+        //                     sessionStorage["user"] = response.data.user;
+        //                     $(".right-panel").hide();
+        //                     // this.$router.push("/home");
+        //                     window.location.href = window.location.href.concat("home");
+        //                 }                        
+        //             })
 
-                    .catch((XMLHttpRequest) => {
-                        throw new Error('Ошибка авторизации', XMLHttpRequest.response.data);
-                    });
-            } 
+        //             .catch((XMLHttpRequest) => {
+        //                 throw new Error('Ошибка авторизации', XMLHttpRequest.response.data);
+        //             });
+        //     } 
             
-            catch (ex) {
-                throw new Error(ex);
-            }
-        },
+        //     catch (ex) {
+        //         throw new Error(ex);
+        //     }
+        // },
 
         onRouteCreateTask() {
             // Если нет роли заказчик, то будет ошибка.
