+++ src/login/login.css
@@ -1,0 +1,451 @@
+body {
+  overflow-x: hidden;
+}
+
+.main-fon {
+  width: 100%;
+}
+
+.block-main-fon {
+  position: relative;
+  text-align: center;
+}
+
+.block-main-fon .title {
+  position: absolute;
+  top: 30%;
+  font-family: Oswald;
+  font-weight: normal;
+  font-size: 40px;
+  color: #000000;
+}
+
+.block-main-fon .title-dop {
+  position: absolute;
+  top: 35%;
+  color: #666666;
+}
+
+.block-main-fon .tablinks:focus {
+  outline: none;
+}
+
+.block-main-fon .btn-create-task {
+  position: absolute;
+  top: 45%;
+  background: #F4511E;
+  color: white;
+  transition: all .3s ease-out;
+  border-radius: initial;
+  border-color: #F4511E;
+  font-family: Oswald;
+  font-size: 12px;
+  border-color: #00BDD6;
+  border: none;
+}
+
+.block-main-fon .btn-create-task:hover {
+  box-shadow: 0 0 10px rgba(0, 0, 0, 0.5);
+}
+
+.block-main-fon .btn-register {
+  position: absolute;
+  top: 45%;
+  background: #00BCD4;
+  color: white;
+  transition: all .3s ease-out;
+  border-radius: initial;
+  border-color: #00BCD4;
+  font-family: Oswald;
+  font-size: 12px;
+}
+
+.block-main-fon .btn-register:hover {
+  box-shadow: 0 0 10px rgba(0, 0, 0, 0.5);
+}
+
+.block-main-fon .centered-fon {
+  display: flex;
+  justify-content: center;
+}
+
+.block-main-fon .left-panel {
+  position: absolute;
+  top: 0;
+  left: -300px;
+}
+
+.block-main-fon .left-panel .block-title {
+  display: flex;
+  justify-content: center;
+}
+
+.block-main-fon .left-panel-not-left {
+  position: absolute;
+  top: 0;
+  padding-left: 0;
+}
+
+.block-main-fon .right-panel {
+  position: absolute;
+  top: 0;
+  right: -20px;
+  width: 318px;
+}
+
+.block-main-fon .right-ul {
+  text-align: start;
+}
+
+.block-main-fon label {
+  margin-bottom: 0;
+}
+
+.block-main-fon label[required=required]:after {
+  color: red;
+  content: " *";
+}
+
+.block-main-fon .asterisk {
+  color: #00BCD4;
+}
+
+.block-main-fon .input-line {
+  display: flex;
+  align-items: center;
+}
+
+.block-main-fon .input-line .input-not-line {
+  background-color: transparent;
+  border: none;
+  border-bottom: 1px solid #b7b7b7;
+  border-radius: 0;
+  outline: 0;
+  height: 20px;
+  width: 100%;
+  font-size: 14px;
+  margin: 0 10px 5px;
+  padding: 0;
+  box-shadow: none;
+  box-sizing: content-box;
+  -webkit-transition: all .3s;
+  transition: all .3s;
+  color: #3c3c3c;
+}
+
+/*.flexcroll {
+  scrollbar-face-color: #367CD2;
+  scrollbar-shadow-color: #FFFFFF;
+  scrollbar-highlight-color: #FFFFFF;
+  scrollbar-3dlight-color: #FFFFFF;
+  scrollbar-darkshadow-color: #FFFFFF;
+  scrollbar-track-color: #FFFFFF;
+  scrollbar-arrow-color: #FFFFFF;
+  height: 100vh;
+}*/
+/* Track */
+/*.flexcroll::-webkit-scrollbar-track {
+  -webkit-box-shadow: inset 0 0 6px rgba(0,0,0,0.3);
+  -webkit-border-radius: 10px;
+  border-radius: 10px;
+}*/
+/* Handle */
+/*.flexcroll::-webkit-scrollbar-thumb {
+  -webkit-border-radius: 10px;
+  border-radius: 10px;
+  background: #00BDD6;
+  -webkit-box-shadow: inset 0 0 6px rgba(0,0,0,0.5);
+}*/
+/* Let's get this party started */
+/*.flexcroll::-webkit-scrollbar {
+  width: 12px;
+}*/
+#idLeftMenu {
+  line-height: 1;
+  color: #98aab7;
+}
+
+#idLeftMenu .left-ul {
+  align-items: flex-start;
+}
+
+#idLeftMenu .left-ul i:hover {
+  color: #00bcd4;
+}
+
+#idLeftMenu .left-ul a:hover {
+  color: #00bcd4;
+}
+
+#idLeftMenu .nav-link {
+  color: #98aab7;
+  cursor: pointer;
+  width: 222px;
+  display: flex;
+}
+
+#idLeftMenu .nav-link:hover {
+  border-left-color: #00bcd4;
+  background-color: #17242f;
+  border-left: 2px solid #00bcd4;
+}
+
+#idLeftMenu .nav-link:hover i {
+  color: #00bcd4;
+}
+
+#idLeftMenu .nav-link:hover a {
+  color: #00bcd4;
+}
+
+#idLeftMenu .nav-link a {
+  color: #98aab7;
+  text-decoration: none;
+}
+
+#idLeftMenu .nav-link a:hover {
+  color: #00bcd4;
+}
+
+#idLeftMenu .icon {
+  font-size: 19px;
+  margin-right: 15px;
+}
+
+#idLeftMenu .icon:hover {
+  color: #00bcd4;
+}
+
+.left-nav {
+  background: #131e27;
+  height: 100vh;
+  line-height: 2;
+}
+
+.right-nav {
+  background: #FFFFFF;
+  height: 100vh;
+  line-height: 2;
+}
+
+.logo-title {
+  color: white;
+  font-size: 24px;
+  text-decoration: none;
+}
+
+.what-is-work {
+  font-family: Open Sans;
+  font-size: 30px;
+  color: #000000;
+}
+
+.content {
+  font-family: Open Sans;
+  font-size: 14px;
+}
+
+.content-title {
+  font-family: Open Sans;
+  font-size: 16px;
+  color: #666666;
+  line-height: 1.5;
+  white-space: pre-line;
+}
+
+.title-color {
+  color: #F4511E;
+}
+
+.gray-fon {
+  background: #FAFAFA;
+}
+
+.card-title {
+  color: #000000;
+  font-family: Open Sans;
+  font-size: 14px;
+}
+
+.card-text {
+  color: #9A9A9A;
+  font-family: Open Sans;
+  font-size: 14px;
+}
+
+/*.card-body {
+  &:hover {
+    box-shadow: 0 0 10px rgba(0, 0, 0, 0.5);
+    transition: all .3s ease-out;
+  }
+}*/
+.title-row {
+  align-items: center;
+  display: flex;
+}
+
+.task-title {
+  color: #00BCD4;
+  font-size: 18px;
+  font-family: Open Sans;
+}
+
+.text-task {
+  color: #444444;
+  font-size: 14px;
+  font-family: Open Sans;
+}
+
+.task-price {
+  color: #00BCD4;
+  font-size: 20px;
+  font-family: Open Sans;
+}
+
+.div-price {
+  margin-left: 15%;
+}
+
+hr {
+  margin: 0;
+}
+
+.task-footer {
+  display: flex;
+  justify-content: space-between;
+}
+
+.task-category {
+  color: #2D7091;
+}
+
+.send-task {
+  color: #444444;
+  font-size: 14px;
+  font-family: Open Sans;
+  color: #00BCD4;
+}
+
+.btn-all-tasks {
+  background: #00BDD6;
+  border-color: #00BDD6;
+  border: none;
+  color: white;
+  border-color: #00BDD6;
+}
+
+.pro {
+  color: #E28327;
+  font-size: 10px;
+}
+
+/* Style the tab */
+.tab {
+  overflow: hidden;
+  background-color: #FFFFFF;
+  display: flex;
+  justify-content: center;
+}
+
+/* Style the buttons that are used to open the tab content */
+.tab button {
+  background-color: inherit;
+  border: none;
+  outline: none;
+  cursor: pointer;
+  padding: 14px 16px;
+  transition: 0.3s;
+  width: 50%;
+}
+
+/* Change background color of buttons on hover */
+.tab button:hover {
+  background-color: #00BDD6;
+  color: black;
+}
+
+/* Create an active/current tablink class */
+.tab button.active {
+  background-color: #00BDD6;
+  color: white;
+}
+
+/* Style the tab content */
+.tabcontent {
+  display: none;
+  padding: 6px 12px;
+  border-top: none;
+}
+
+.tab-role {
+  background: white;
+}
+
+/* Change background color of buttons on hover */
+.tab-role button:hover {
+  background-color: #00BDD6;
+  color: black;
+}
+
+.tab-role button:active {
+  background-color: #00BDD6;
+}
+
+/* Create an active/current tablink class */
+.tab-role {
+  background-color: white;
+}
+
+.role-hide {
+  display: none;
+}
+
+.role-show {
+  display: block;
+}
+
+#idRightMenu {
+  justify-content: center;
+}
+
+.text-role {
+  font-family: Open Sans;
+}
+
+.selected-role {
+  background: #00BCD4;
+}
+
+.not-selected-role {
+  background: white;
+}
+
+.block-work {
+  flex-wrap: wrap;
+}
+
+.form-register {
+  width: 30%;
+  background: #00BDD6;
+  color: white;
+  transition: all .3s ease-out;
+  border-radius: initial;
+  font-family: Oswald;
+  border: none;
+}
+
+.form-register :focus {
+  outline: none;
+}
+
+.btn-register {
+  background: #f4511e;
+  color: #fff;
+  transition: all .3s ease-out;
+  border-radius: initial;
+  border-color: #f4511e;
+  font-family: Oswald;
+  font-size: 12px;
+  border-color: #00bdd6;
+  border: none;
+}
