+++ src/login/login.html
@@ -11,18 +11,18 @@
     <div class="mt-3 d-flex justify-content-center">
         <div class="form-register">
             <div>
-                <label>Login или E-mail</label><strong class="asterisk">&#42;</strong>
-                <input type="text" class="form-control" aria-label="First name">
+                <label for="idEma">Login или E-mail</label><strong class="asterisk">&#42;</strong>
+                <input type="text" id="idEma" class="form-control" aria-label="First name">
             </div>        
 
             <div class="mt-3">
                 <label>Пароль</label><strong class="asterisk">&#42;</strong>
-                <input type="text" class="form-control" aria-label="First name">
+                <input id="idPass" v-model="sPassword" type="text" class="form-control" aria-label="First name">
             </div>     
 
             <div class="mt-3 text-center">
                 <button class="btn-register form-control-sm btn-create-task btn-all-tasks mb-3"
-                      _v-on:click="onCreate()">
+                      v-on:click="onLogin()">
                         ВОЙТИ
                 </button>
             </div>
